+++ app/services/forest_liana/base_getter.rb
@@ -1,0 +1,12 @@
+module ForestLiana
+  class BaseGetter
+    def get_resource
+      use_act_as_paranoid = @resource.instance_methods
+        .include? :really_destroyed?
+
+      # NOTICE: Do not unscope with the paranoia gem to prevent the retrieval
+      #         of deleted records.
+      use_act_as_paranoid ? @resource : @resource.unscoped
+    end
+  end
+end