+++ app/services/forest_liana/stat_getter.rb
@@ -1,5 +1,5 @@
 module ForestLiana
-  class StatGetter
+  class StatGetter < BaseGetter
     attr_accessor :record
 
     def initialize(resource, params)